fix moto build.prop

+++ tools/system_vendor_prop.sh
@@ -14,7 +14,7 @@ source $PROJECT_DIR/helpers/common_script.sh "y"
 
 # Exit if no arguements
 if [ -z "$1" ] ; then
-    echo -e "Supply sytem &/ vendor build.prop as arguements!"
+    echo -e "Supply ROM directory as arguement!"
     exit 1
 fi
 
@@ -28,7 +28,10 @@ elif [ -d "$1" ]; then
         SYSTEM_PATH="system"
     fi
     find "$1/$SYSTEM_PATH" -maxdepth 1 -name "build*prop" -exec cat {} >> $PROJECT_DIR/working/system_working.prop \;
-    find "$1/vendor" -maxdepth 1 -name "build*prop" -exec cat {} >> $PROJECT_DIR/working/vendor_working.prop \;
+    [[ -d "$1/system/vendor/" ]] && VENDOR_PATH="system/vendor"
+    [[ -d "$1/system/system/vendor/" ]] && VENDOR_PATH="system/system/vendor"
+    [[ -d "$1/vendor/" ]] && VENDOR_PATH="vendor"
+    find "$1/$VENDOR_PATH" -maxdepth 1 -name "build*prop" -exec cat {} >> $PROJECT_DIR/working/vendor_working.prop \;
 else
     cp -a $1 $PROJECT_DIR/working/system_working.prop
 fi
@@ -41,16 +44,18 @@ if [ ! -z "$2" ] ; then
 fi
 
 # system.prop
-TSTART=$(grep -nr "# end build properties" $PROJECT_DIR/working/system_working.prop | sed "s|:.*||g")
-TSTART=$((TSTART+1))
-TEND=$(grep -nr "# ADDITIONAL_BUILD_PROPERTIES" $PROJECT_DIR/working/system_working.prop | sed "s|:.*||g")
-TEND=$((TEND-1))
-sed -n "${TSTART},${TEND}p" $PROJECT_DIR/working/system_working.prop > $PROJECT_DIR/working/system.prop
+if [ -s "$PROJECT_DIR/working/system_working.prop" ]; then
+    TSTART=$(grep -nr "# end build properties" $PROJECT_DIR/working/system_working.prop | sed "s|:.*||g" | head -1)
+    TSTART=$((TSTART+1))
+    TEND=$(grep -nr "# ADDITIONAL_BUILD_PROPERTIES" $PROJECT_DIR/working/system_working.prop | sed "s|:.*||g" | head -1)
+    TEND=$((TEND-1))
+    sed -n "${TSTART},${TEND}p" $PROJECT_DIR/working/system_working.prop > $PROJECT_DIR/working/system.prop
+fi
 
 # vendor.prop
-if [ ! -z "$2" ] || [ -d "$1" ]; then
-    TSTART=$(grep -nr "ADDITIONAL VENDOR BUILD PROPERTIES" $PROJECT_DIR/working/vendor_working.prop | sed "s|:.*||g")
-    TEND=$(wc -l $PROJECT_DIR/working/vendor_working.prop | sed "s| .*||g")
+if [ -s "$PROJECT_DIR/working/vendor_working.prop" ]; then
+    TSTART=$(grep -nr "ADDITIONAL VENDOR BUILD PROPERTIES" $PROJECT_DIR/working/vendor_working.prop | sed "s|:.*||g" | head -1)
+    TEND=$(wc -l $PROJECT_DIR/working/vendor_working.prop | sed "s| .*||g" | head -1)
     sed -n "${TSTART},${TEND}p" $PROJECT_DIR/working/vendor_working.prop | sort | sed "s|#.*||g" | sed '/^[[:space:]]*$/d' > $PROJECT_DIR/working/vendor_new.prop
     sed -i -e 's/^/    /' $PROJECT_DIR/working/vendor_new.prop
     sed -i '1 i\PRODUCT_PROPERTY_OVERRIDES += ' $PROJECT_DIR/working/vendor_new.prop
